Fix SettingsClient import and component reference in page.tsx
