+++ apps/web/src/app/SettingsPage/components/ClientOnly.tsx
@@ -1,0 +1,21 @@
+'use client';
+
+import { useEffect, useState, ReactNode } from 'react';
+
+interface ClientOnlyProps {
+  children: ReactNode;
+}
+
+export default function ClientOnly({ children }: ClientOnlyProps) {
+  const [hasMounted, setHasMounted] = useState(false);
+
+  useEffect(() => {
+    setHasMounted(true);
+  }, []);
+
+  if (!hasMounted) {
+    return null;
+  }
+
+  return <>{children}</>;
+}
