+++ apps/web/src/app/SettingsPage/components/SettingsClient.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+import { FC } from "react";
+import SettingsContent from "./SettingsContent";
+import ClientOnly from "./ClientOnly";
+
+const SettingsClient: FC = () => {
+  return (
+    <ClientOnly>
+      <SettingsContent />
+    </ClientOnly>
+  );
+};
+
+export default SettingsClient;
